// src/app/page/master/master-pegawai-sk-fasilitas/master-pegawai-sk-fasilitas.interface.ts
import {EventEmitter} from '@angular/core';

export interface MasterPegawaiSkFasilitas {
	
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
 }